Move tarball to tarball.server
